refactor: use a wrapper for checkMatrix
