+++ src/decompositions.js
@@ -1,11 +1,11 @@
-import Matrix from './matrix';
+import {Matrix, WrapperMatrix2D} from '..';
 
 import LuDecomposition from './dc/lu';
 import QrDecomposition from './dc/qr';
 import SingularValueDecomposition from './dc/svd';
 
 export function inverse(matrix, useSVD = false) {
-    matrix = Matrix.checkMatrix(matrix);
+    matrix = WrapperMatrix2D.checkMatrix(matrix);
     if (useSVD) {
         return SingularValueDecomposition(matrix).inverse();
     } else {
@@ -14,8 +14,8 @@ export function inverse(matrix, useSVD = false) {
 }
 
 export function solve(leftHandSide, rightHandSide, useSVD = false) {
-    leftHandSide = Matrix.checkMatrix(leftHandSide);
-    rightHandSide = Matrix.checkMatrix(rightHandSide);
+    leftHandSide = WrapperMatrix2D.checkMatrix(leftHandSide);
+    rightHandSide = WrapperMatrix2D.checkMatrix(rightHandSide);
     if (useSVD) {
         return SingularValueDecomposition(leftHandSide).solve(rightHandSide);
     } else {
